create PlayerController
