+++ Assets/Scripts/GameManager.cs
@@ -12,10 +12,7 @@ public class GameManager : MonoBehaviour
 	[SerializeField] FoodGenerater foodGenerater;
 	[SerializeField] ItemGenerater itemGenerater;
 	[SerializeField] LineCursol lineCursol;
-	[SerializeField] Text scoreText;
-	[SerializeField] Text remainLinesText;
-	[SerializeField] Text comboText;
-	[SerializeField] Image rushGageImage;
+	[SerializeField] PlayerController playerController;
 	[SerializeField] Image calorieGageImage;
 
 	[System.NonSerialized] public List<Oil> Oils;
@@ -25,7 +22,7 @@ public class GameManager : MonoBehaviour
 	[System.NonSerialized] public HorizontalLine[,] AmidaLines;
 
 	bool isRush = false;
-	bool IsRush
+	public bool IsRush
 	{
 		get { return isRush; }
 
@@ -44,95 +41,6 @@ public class GameManager : MonoBehaviour
 		}
 	}
 
-	//●Playerクラスに分ける可能性あり●
-	const int MaxDrawLineNum = 3;
-
-	const float RushTime = 20;
-	const int MaxRushGage = 1;
-	int rushGage;
-	public int RushGage
-	{
-		get { return rushGage; }
-
-		set
-		{
-			if (IsRush)
-			{
-				return;
-			}
-
-			if (value >= MaxRushGage)
-			{
-				rushGage = MaxRushGage;
-			}
-			else
-			{
-				rushGage = value;
-			}
-
-			if (rushGage >= MaxRushGage)
-			{
-				Debug.Log("RushTime");
-				StartCoroutine(RushTimeCoroutine());
-			}
-		}
-	}
-	int score;
-	public int Score
-	{
-		get { return score; }
-
-		set
-		{
-			score = value;
-			scoreText.text = "SCORE:" + Score.ToString();
-		}
-	}
-
-	int remainLines = MaxDrawLineNum;
-	public int RemainLines
-	{
-		get { return remainLines; }
-
-		set
-		{
-			if (value > MaxDrawLineNum)
-			{
-				remainLines = MaxDrawLineNum;
-			}
-			else if (value < 0)
-			{
-				remainLines = 0;
-			}
-			else
-			{
-				remainLines = value;
-			}
-			remainLinesText.text = "残り本数" + RemainLines.ToString() + "/" + MaxDrawLineNum.ToString();
-		}
-	}
-
-	int combo = 0;
-	public int Combo
-	{
-		get { return combo; }
-
-		set
-		{
-			combo = value;
-			if (combo == 0)
-			{
-				comboText.gameObject.SetActive(false);
-			}
-			else
-			{
-				comboText.gameObject.SetActive(true);
-				comboText.text = combo.ToString() + "コンボ";
-			}
-		}
-	}
-	int maxCombo;
-
 	// Start is called before the first frame update
 	void Start()
     {
@@ -145,6 +53,19 @@ public class GameManager : MonoBehaviour
 		{
 			trash.completedFriedFoodDelegate = CompletedFriedFood;
 		}
+
+		playerController.rushDelegate = (bool isRush) =>
+		{
+			itemGenerater.IsRush = isRush;
+			//●もっときれいにやりたい●
+			foreach (Trash trash in Trashes)
+			{
+				trash.ChangeOil(isRush);
+			}
+			//BGM変更
+			//ゲーム速度少し上昇
+			//制限時間停止
+		};
 
 		Invoke("GameStart", 0.5f);
 	}
@@ -214,21 +135,11 @@ public class GameManager : MonoBehaviour
 			currentCustomer.CustomerReact(friedFood,
 				(int rushGage, int score) =>
 				{
-					Score += IsRush ? (int)(score * 1.5f) : score;
-					RushGage += rushGage;
-					Debug.Log(score);
+					playerController.Score += playerController.IsRush ? (int)(score * 1.5f) : score;
+					playerController.RushGage += rushGage;
 				});
 
-			//コンボ処理
-			if (friedFood.FriedFoodReview == Cooking.FriedFoodReview.good)
-			{
-				Combo++;
-			}
-			else
-			{
-				maxCombo = Math.Max(maxCombo, Combo);
-				Combo = 0;
-			}
+			playerController.Combo = friedFood.FriedFoodReview == Cooking.FriedFoodReview.good ? playerController.Combo + 1 : 0;
 		}
 		else
 		{
@@ -251,27 +162,5 @@ public class GameManager : MonoBehaviour
 
 		//9：次のステージへ
 		Debug.Log("clear");
-	}
-
-	void RushEnd()
-	{
-		IsRush = false;
-	}
-
-	IEnumerator RushTimeCoroutine()
-	{
-		float time = 0;
-		IsRush = true;
-
-		while (time < RushTime)
-		{
-			time += Time.deltaTime;
-			rushGageImage.fillAmount = (RushTime - time) / RushTime;
-			yield return null;
-		}
-
-		rushGage = 0;
-		rushGageImage.fillAmount = 1;
-		IsRush = false;
 	}
 }
